First step in moving to kebap menu
