+++ app/src/main/res/values-de/strings.xml
@@ -72,4 +72,5 @@
     <string name="time_after">x Stunden nach Medikament y</string>
     <string name="instruction_templates">Hinweisvorlagen</string>
     <string name="instructions_label">Zusätzliche Hinweise (z.B. vor einer Mahlzeit)</string>
+    <string name="backup_medicine">Backup der Medikamente</string>
 </resources>
