+++ app/src/main/res/values-es/strings.xml
@@ -71,4 +71,5 @@
     <string name="time_after">x horas después del medicamento y</string>
     <string name="instruction_templates">Plantillas de instrucciones</string>
     <string name="instructions_label">Instrucciones adicionales (por ejemplo, antes de las comidas)</string>
+    <string name="backup_medicine">Backup de los medicamentos</string>
 </resources>
