+++ app/src/main/res/values/arrays.xml
@@ -29,10 +29,6 @@
         <item>48</item>
         <item>72</item>
     </string-array>
-    <string-array name="export_formats">
-        <item>CSV</item>
-        <item>PDF</item>
-    </string-array>
     <string-array name="instructions_suggestions">
         <item>@string/no_instructions</item>
         <item>@string/before_meal</item>
