+++ app/src/main/res/values/strings.xml
@@ -73,4 +73,7 @@
     <string name="time_after">x hours after medicine y</string>
     <string name="instruction_templates">Instruction templates</string>
     <string name="instructions_label">Additional instructions (e.g. before meals)</string>
+    <string name="backup_medicine">Backup medicine settings</string>
+    <string name="csv" translatable="false">CSV</string>
+    <string name="pdf" translatable="false">PDF</string>
 </resources>
