+++ app/src/main/res/xml/root_preferences.xml
@@ -66,6 +66,11 @@
         app:key="generate_test_data"
         app:summary=""
         app:title="@string/generate_test_data" />
+    <Preference
+        app:icon="@drawable/x_circle"
+        app:key="generate_test_data"
+        app:summary=""
+        app:title="@string/generate_test_data" />
     <PreferenceCategory
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
